refactor: TestConfig를 공통으로 사용하도록 수정
