+++ metanet4/src/test/java/com/metanet/team4/payment/controller/PaymentControllerTest.java
@@ -22,6 +22,7 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.metanet.team4.common.CommonLoginTestConfig;
 import com.metanet.team4.common.Login;
 import com.metanet.team4.jwt.JwtAuthenticationFilter;
 import com.metanet.team4.member.model.Member;
@@ -32,7 +33,7 @@ import com.metanet.team4.payment.service.PaymentService;
 @ExtendWith(SpringExtension.class)
 @SpringBootTest
 @AutoConfigureMockMvc
-@Import(PaymentControllerTest.TestConfig.class)  // 커스텀 ArgumentResolver 등록을 원한다면 유지
+@Import(CommonLoginTestConfig.class)
 class PaymentControllerTest {
 
     @MockBean
@@ -46,37 +47,6 @@ class PaymentControllerTest {
 
     @Autowired
     private ObjectMapper objectMapper;
-
-    @TestConfiguration
-    static class TestConfig {
-
-        /** 
-         * MockMvc 테스트 시, @Login 파라미터를 강제 주입하기 위한 커스텀 ArgumentResolver 예시
-         * 원본 코드를 유지하려면 아래처럼 Bean 등록하거나,
-         * 또는 WebMvcConfigurer를 구현해서 addArgumentResolvers() 등록하는 식으로 바꿀 수도 있음
-         */
-        @Bean
-        public HandlerMethodArgumentResolver loginArgumentResolver() {
-            return new HandlerMethodArgumentResolver() {
-                @Override
-                public boolean supportsParameter(MethodParameter parameter) {
-                    return parameter.hasParameterAnnotation(Login.class);
-                }
-                @Override
-                public Object resolveArgument(MethodParameter parameter,
-                                              ModelAndViewContainer mavContainer,
-                                              org.springframework.web.context.request.NativeWebRequest webRequest,
-                                              WebDataBinderFactory binderFactory) {
-                    // 테스트용 Member 목 객체 생성
-                    Member mockMember = new Member();
-                    mockMember.setId(999L);
-                    mockMember.setName("MockUser");
-                    // 필요하면 다른 필드도 세팅
-                    return mockMember;
-                }
-            };
-        }
-    }
 
     @Test
     void requestPaymentTest() throws Exception {
