+++ metanet4/src/test/java/com/metanet/team4/reservation/controller/ReservControllerTest.java
@@ -11,22 +11,14 @@ import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
-import org.springframework.core.MethodParameter;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.metanet.team4.common.Login;
+import com.metanet.team4.common.CommonLoginTestConfig;
 import com.metanet.team4.jwt.JwtAuthenticationFilter;
-import com.metanet.team4.member.model.Member;
 import com.metanet.team4.payment.model.CancelResponseDto;
 import com.metanet.team4.payment.model.ReservationDetailDto;
 import com.metanet.team4.payment.service.ReservService; 
@@ -34,7 +26,7 @@ import com.metanet.team4.payment.service.ReservService;
 @ExtendWith(SpringExtension.class)
 @SpringBootTest
 @AutoConfigureMockMvc
-@Import(ReservControllerTest.TestConfig.class)
+@Import(CommonLoginTestConfig.class)
 class ReservControllerTest {
 
     @MockBean
@@ -48,30 +40,6 @@ class ReservControllerTest {
 
     @Autowired
     private ObjectMapper objectMapper;
-
-    @TestConfiguration
-    static class TestConfig {
-        @Bean
-        public HandlerMethodArgumentResolver loginArgumentResolver() {
-            return new HandlerMethodArgumentResolver() {
-                @Override
-                public boolean supportsParameter(MethodParameter parameter) {
-                    return parameter.hasParameterAnnotation(Login.class);
-                }
-                @Override
-                public Object resolveArgument(MethodParameter parameter,
-                                              ModelAndViewContainer mavContainer,
-                                              NativeWebRequest webRequest,
-                                              WebDataBinderFactory binderFactory) {
-                    // 테스트용 Member 목 객체
-                    Member mockMember = new Member();
-                    mockMember.setId(999L);
-                    mockMember.setName("MockUser");
-                    return mockMember;
-                }
-            };
-        }
-    }
 
     @Test
     void getReservationDetailTest() throws Exception {
